sun5i: Give PLL6 a sane initial value
The upstream kernel does not override PLL values as set by power-on-reset
and/or the bootloader if it can avoid doing so.

The power-on-reset value for PLL6 is 2400MHz, which does not actually work,
the kernel does not know this and 2400 / 48 is 50 MHz, so the kernel happily
tries to use it as is for mmc.

Set PLL6 to a working frequency so that the kernel can use it. This fixes
mmc in the upstream kernel not working on sun5i.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>