sunxi: fast MBUS support on sun7i
For qualifying hardware, we can run MBUS at 400MHz without risking
stability if we raise the DCDC3 voltage to 1.3V .

This is desirable since on A20 the MBUS is a significant bottle-neck
for some workloads.

This has been tested on 50 pieces of Cubieboard2 by Cubietech and no
issues of any kind were reported.

Signed-off-by: Emilio López <emilio@elopez.com.ar>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>